+++ apps/frontend/src/modules/surveys/components/studio/create/editor-panel/index.tsx
@@ -2,8 +2,8 @@ import { HelpCircle, SettingsIcon } from "@/ui/icons";
 import * as Tabs from "@radix-ui/react-tabs";
 import * as Tooltip from "@radix-ui/react-tooltip";
 import { Pen } from "lucide-react";
-import UpdateQuestion from "../index.tsx";
 import React from "react";
+import UpdateQuestion from "../index.tsx";
 import { Preview } from "../preview-panel/index.tsx";
 import { UpdateDesignEditor } from "./DesignEditor";
 import { UpdateSettingsEditor } from "./SettingsEditor";
@@ -40,13 +40,13 @@ export default function Create() {
                     <Tabs.Trigger
                         value={label}
                         key={index}
-                        className="data-[state=active]:bg-intg-bg-15 data-[state=active]:rounded-md"
+                        className="data-[state=active]:rounded-md data-[state=active]:bg-intg-bg-15"
                         asChild
                     >
                         <div>
                             <Tooltip.Provider>
                                 <Tooltip.Root key={id}>
-                                    <Tooltip.Trigger className="hover:bg-intg-bg-15 h-9 rounded px-2 ease-in-out hover:transition-all">
+                                    <Tooltip.Trigger className="h-9 rounded px-2 ease-in-out hover:bg-intg-bg-15 hover:transition-all">
                                         {icon}
                                     </Tooltip.Trigger>
