+++ apps/frontend/src/modules/surveys/components/studio/create/editor-panel/questions/index.tsx
@@ -1,6 +1,6 @@
 import { addEllipsis, cn } from "@/utils";
 import * as Accordion from "@radix-ui/react-accordion";
-import { useEffect, useMemo, useState } from "react";
+import { useMemo, useState } from "react";
 import { QuestionOptions } from "./attributes/Options";
 
 import { SurveyQuestionTypeEnum } from "@/generated/graphql";
@@ -9,8 +9,7 @@ import { questionTypes } from "@/utils/survey";
 import { QuestionPanel } from "./QuestionPanel";
 
 export default function UpdateQuestion() {
-    const { getSurvey, questions, setOpenQuestion, openQuestion, surveySlug } =
-        useSurvey();
+    const { questions, setOpenQuestion, openQuestion } = useSurvey();
 
     const [currentQuestionType, setCurrentQuestionType] = useState<
         SurveyQuestionTypeEnum | undefined
@@ -22,7 +21,6 @@ export default function UpdateQuestion() {
         });
     }, [questions]);
 
- 
     return (
         <div className="h-full w-full space-y-4 pt-2">
             <div>
