+++ apps/frontend/src/pages/SurveyStudio.tsx
@@ -1,5 +1,10 @@
+import { SurveyProvider } from "@/modules/surveys/SurveyProvider";
 import Studio from "@/modules/surveys/components/Studio";
 
 export function SurveyStudio() {
-    return <Studio />;
+    return (
+        <SurveyProvider>
+            <Studio />
+        </SurveyProvider>
+    );
 }
